refactor: standardize abstract-base file layout (_base.py → base.py, media package)
Renames `_base.py` → `base.py` in catalog and provenance models packages
to match Django's own convention (e.g. `django/db/models/base.py`). The
underscore prefix was a Pinbase-only convention.

Converts `apps/media/models.py` to a `models/` package (`asset.py`,
`rendition.py`, `entity.py`, plus a re-exporting `__init__.py`) so media
follows the same layout as catalog and provenance. This is the substrate
for Step 9.2 of the mypy burn-down, which adds an abstract base
(`MediaSupported`) to the media app.

Pure structural refactor: no behavior change, no new abstractions,
no migrations. Existing `from apps.media.models import …` import paths
continue to work via package re-exports.
