+++ backend/apps/media/models/__init__.py
@@ -1,0 +1,11 @@
+"""Media models: storage infrastructure and catalog attachment.
+
+Re-exports all public names so existing ``from apps.media.models import …``
+continues to work unchanged.
+"""
+
+from .asset import MediaAsset
+from .entity import EntityMedia
+from .rendition import MediaRendition
+
+__all__ = ["EntityMedia", "MediaAsset", "MediaRendition"]
